Implement a simple routing policy.
Implement a simple set of routing policies for audio outputs; either
all plugged outputs, or the most recently plugged output.  Eventually,
this should go away and be replaced by more sophisticated policy
management don't by a different (replaceable) policy service, but for
now it is nice to have some functionality here, even if it is limited.

Change-Id: Id5ad4c11d59592283369a1ee012bced9eccfb62e