apps/media: remove rendundant flog message
Packet producers log both a 'Connecting' and a 'ConnectedTo' flog
message. These are redundant, ConnectedTo being the more
expressive message. This CL gets rid of Connecting. It also moves
the ConnectedTo call from fidl_packet_producer to the base class
media_packet_producer_base, which does all the rest of the
producer flogging.

The 'digest' analyzer now properly disconnects producer and consumer
when the 'Reset' message is processed. This is correct but a bit
annoying, because we'll often see disassembled graphs when we process
an entire file. This can be addressed using the --stop-index option,
but it would great to have more fluid tools that made this easier.

Change-Id: I8bc0b9a99f2debd2e73da54ef593877e25fe4b77