apps/media: combine flog full/digest channel handlers
I was using different channel handler classes for full/terse format
and digest format. This makes for a lot more files and discourages
me from adding new channels. This CL merges the handlers together so
each one handles all three formats for a particular channel type.
We can still add new format types if we want, but that's not likely
to happen.

Change-Id: I3b5484e2a43c89bceed39b2147d87c78e5ac4c55