apps/media: clean up in connection error handlers
This CL adds cleanup code in connection error handlers (the functions
passed to set_connection_error_handler for bindings and interface
pointers). In all cases, we close/reset the binding/interface in
question. In the case of bindings, we clear the connection error
handler. It appears that InterfacePtr takes care of this itself when
reset.

It's doubtful that any of these changes fix real leaks, because the
objects that contain the bindings and interface pointers in question
are typically destroyed in short order. On the other hand, it's
probably good practice to be proactive about cleanup and to locate
the cleanup where the mess is created rather than guessing that it's
not necessary.

Change-Id: I2488e2c4e8e17ab40713ac75bfdd656d367c61fc