apps/media: don't fault on seek out of range (MTWN-11)
Ffmpeg is requesting out-of-range seeks, which cause a DCHECK. This
CL replaces the DCHECK with a DLOG(ERROR), so we'll still see the
message in debug builds, but it won't blow up the media service. The
seek function returns EIO when this happens. If this causes problems,
we'll see the message and correct from there.

Change-Id: I8ce97df171a135024626a7ad21839eab417d04ff