apps/media: expose show/hide control overlay logic
In the dart/flutter libs, the logic as to whether to show the control
overlay was implemented in the media player widget and not exposed.
That worked, but we really want this logic to be available to client
code so that additional overlay UI with the same show/hide behavior
can be added.

It's easier for client code to access the controller vs the widget,
so this CL moves the logic to MediaPlayerController. Control overlay
is irrelevant to AudioPlayerController, so it's not implemented
there.

MediaPlayerController now exposes a shouldShowControlOverlay read-
only property, a brieflyShowControlOverlay method and a
overlayAutoHideDuration read/write property. Listeners of the
controller are notified when shouldShowControlOverlay changes.

Change-Id: I388554a1808e5e98ebad4ed6b171a3e764df49cf