apps/media: fix player hang after seek
Sometimes after a seek, the player would hang waiting for the
renderers to finish priming. This was happening because of a race
condition which caused stale demand to be delivered to the demux.
This caused one of the streams out of the demux to get stuck
thinking there was no demand for packets. Consequently, the
renderer on that stream would never get the packets it needed to
finish priming, and the player would wait forever for priming to
complete.

The race condition in question isn't really resolved, so this
fix qualifies as a workaround. For the record, the race occurs
because the demand returned by FidlPacketProducer::SupplyPacket
arrives *after* the demand callback is called in
FidlPacketProducer::SendPacket. By making all this happen on
the same thread, we avoid the race.

Change-Id: I7d2ba052b1cfb0db641be1b7fb83144406a38c09